+++ package.json
@@ -13,13 +13,10 @@
     "check-format": "prettier --check .",
     "check-lint": "eslint . --ext ts --ext tsx --ext js",
     "reload": "pm2 startOrGracefulReload ecosystem.config.js",
-    "cypress": "cypress open",
-    "cypress:headless": "cypress run --config video=false",
-    "test:component": "cypress run-ct --record",
     "screenshots": "cypress run --env isTest=false --spec 'cypress/e2e/screenshots/*.cy.ts' --config baseUrl='https://interfacer-gui-staging.dyne.org/'",
     "e2e": "pnpm build && start-server-and-test start http://localhost:3000 cypress",
     "e2e:headless": "pnpm build && start-server-and-test start http://localhost:3000 cypress:headless",
-    "test": "pnpm build && start-server-and-test start http://localhost:3000 cypress:headless",
+    "test": "pnpm exec playwright test",
     "translate": "./bin/translate.sh",
     "prepare": "husky install",
     "release": "HUSKY=0 standard-version --no-verify",
