+++ playwright.config.ts
@@ -1,25 +1,20 @@
 import { defineConfig, devices } from "@playwright/test";
 import dotenv from "dotenv";
 
-/**
- * Read environment variables from file.
- * https://github.com/motdotla/dotenv
- */
-
-/**
- * See https://playwright.dev/docs/test-configuration.
- */
-
 dotenv.config({ path: "./playwright.env" });
 
 export default defineConfig({
+  webServer: {
+    command: "pnpm build && pnpm start",
+    url: "http://localhost:3000",
+  },
   testDir: "./tests",
   /* Run tests in files in parallel */
   fullyParallel: true,
   /* Fail the build on CI if you accidentally left test.only in the source code. */
   forbidOnly: !!process.env.CI,
   /* Retry on CI only */
-  retries: process.env.CI ? 2 : 0,
+  retries: 0,
   /* Opt out of parallel tests on CI. */
   workers: process.env.CI ? 1 : undefined,
   /* Reporter to use. See https://playwright.dev/docs/test-reporters */
